Remove wordwrap for links so they don't break (on mobile and smaller screen sizes)
The current implementation of wordwrap is only used in the case of a
link. At which point it checks against a maxWidth of 76 and breaks the
link, and/or inserts a newline character if whitespace is found.

Introducing a newline character breaks the link and is seemingly unnecessary
(may have to do with addonFit).

Introducing a newline on whitespace character is also seemingly
unecessary as a URL should not have a whitespace character anywhere but
possibly the end of the string.