Print email, subject, and body of apply mailto:
Currently the apply command line feature is broken because we are programmatically navigating to a mailto link which is blocked in current browsers (see: https://stackoverflow.com/questions/71583658/not-allowed-to-launch-cutom-protocol-because-a-user-gesture-is-required.)

I first try to rewrite the `term.stylePrint` with a new mailto: regex in addition to the http regex. My next approach was to write a new `term.displayMailto` function.

Both of these failed because of the poor support for mailto in current browsers.

This is a temporary fix so that users know that apply isn't broken.

In the future if you'd make a form endpoint available to me I'd love to write an interactive application prompt where they could answer a few questions right there and have the mail send via our backend.